requirements.txt
requirements-parser
requests
pyfiglet
python-whois